requirements.txt
keras==2.15.0
matplotlib==3.8.2
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.2.0
scikit_learn==1.3.2
streamlit==1.32.2
tensorflow==2.15.0

